Refactor: 상대경로 수정, rental, group, reservation page 구조 변경
