+++ src/shared/routes/routes-config.ts
@@ -15,7 +15,7 @@ export const ROUTES = {
   SCAN: '/scan',
   CART: '/cart',
   NOTIFICATION: '/notification',
-  MY_RESERVATION: '/mypage/reservation',
-  MY_LOAN: '/mypage/loan',
-  MY_GROUP: '/mypage/group',
+  RESERVATION: '/reservation',
+  RENTAL: '/rental',
+  GROUP: '/group',
 };
